Safari-611.1.10/LayoutTests/media/track/in-band/track-in-band-srt-mkv-language.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../../media-file.js></script>
        <script src=../../video-test.js></script>
        <script src=../../in-band-tracks.js></script>
    </head>
    <body onload="testAttribute('../../content/counting-subtitled-srt.mkv', 'text', 'language', ['', 'en', 'fr'])">
        <video controls></video>
        <p>Check in-band text tracks' language attributes.</p>
    </body>
</html>
