Safari-611.1.10/LayoutTests/media/track/in-band/track-in-band-srt-mkv-language-expected.txt


Check in-band text tracks' language attributes.

EVENT(canplaythrough)

** Check in-band kind attributes
EXPECTED (video.textTracks.length == '3') OK
EXPECTED (video.textTracks[0].language == '') OK
EXPECTED (video.textTracks[1].language == 'en') OK
EXPECTED (video.textTracks[2].language == 'fr') OK

END OF TEST
